Fix CI tests
